Address comments by @minwoox, add NoopKeepAliveHandler
