+++ core/src/main/java/com/linecorp/armeria/internal/common/Http1ObjectEncoder.java
@@ -211,9 +211,7 @@ public abstract class Http1ObjectEncoder implements HttpObjectEncoder {
             final ChannelFuture future = ch.write(obj);
             if (!isPing(id)) {
                 final KeepAliveHandler keepAliveHandler = keepAliveHandler();
-                if (keepAliveHandler != null) {
-                    keepAliveHandler.onReadOrWrite();
-                }
+                   keepAliveHandler.onReadOrWrite();
             }
             if (endStream) {
                 currentId++;
@@ -341,10 +339,7 @@ public abstract class Http1ObjectEncoder implements HttpObjectEncoder {
         }
         closed = true;
 
-        final KeepAliveHandler keepAliveHandler = keepAliveHandler();
-        if (keepAliveHandler != null) {
-            keepAliveHandler.destroy();
-        }
+        keepAliveHandler().destroy();
         if (pendingWritesMap.isEmpty()) {
             return;
         }
