+++ core/src/main/java/com/linecorp/armeria/internal/common/KeepAliveHandlerUtil.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2021 LINE Corporation
+ *
+ * LINE Corporation licenses this file to you under the Apache License,
+ * version 2.0 (the "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at:
+ *
+ *   https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+
+package com.linecorp.armeria.internal.common;
+
+public final class KeepAliveHandlerUtil {
+
+    public static boolean needKeepAliveHandler(long idleTimeoutMillis, long pingIntervalMillis,
+                                               long maxConnectionAgeMillis, int maxNumRequests) {
+        return idleTimeoutMillis > 0 || pingIntervalMillis > 0 ||
+               maxConnectionAgeMillis > 0 || maxNumRequests > 0;
+    }
+
+    private KeepAliveHandlerUtil() {}
+}
