+++ core/src/test/java/com/linecorp/armeria/client/HttpResponseWrapperTest.java
@@ -32,6 +32,7 @@ import com.linecorp.armeria.common.ResponseHeaders;
 import com.linecorp.armeria.common.logging.RequestLogProperty;
 import com.linecorp.armeria.internal.common.InboundTrafficController;
 import com.linecorp.armeria.internal.common.KeepAliveHandler;
+import com.linecorp.armeria.internal.common.NoopKeepAliveHandler;
 
 import io.netty.channel.Channel;
 import reactor.test.StepVerifier;
@@ -169,7 +170,7 @@ class HttpResponseWrapperTest {
 
         @Override
         KeepAliveHandler keepAliveHandler() {
-            return null;
+            return NoopKeepAliveHandler.INSTANCE;
         }
     }
 }
